WIP: add receive location form action & clean up
